pyproject.toml
[tool.poetry]
name = "tarly"
version = "0.1.0"
description = "Backend test exam"
authors = ["sergeibersadsky <sergeibershadsky@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.2"
django-extensions = "^2.1"
djangorestframework = "^3.9"
model_mommy = "^1.6"
django-filter = "^2.1"
python-dotenv = "^0.10.2"
psycopg2_binary = "^2.8"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
